+++ graphannis/src/annis/db/corpusstorage.rs
@@ -1992,6 +1992,20 @@ impl CorpusStorage {
     /// - `node_ids` - A set of node annotation identifiers describing the subgraph.
     /// - `ctx_left` and `ctx_right` - Left and right context in token distance to be included in the subgraph.
     /// - `segmentation` - The name of the segmentation which should be used to as base for the context. Use `None` to define the context in the default token layer.
+    ///
+    /// ## Changes to the subgraph in comparision to the original graph
+    ///
+    /// The context definition can cause gaps in the returned subgraph, e.g. if
+    /// the given node ids are too far apart for their context to overlap. Since
+    /// only edges for the nodes of the contexts are included, it is impossible
+    /// use the original Ordering edges to sort the results alone, since there will be no
+    /// connection between the tokens of the non-overlapping context regions.
+    ///
+    /// To allow sorting the non-overlapping context regions by their order in
+    /// the datasource, an edge in the special `Ordering/annis/datasource-gap`
+    /// component is added between the last token of each context region and the
+    /// first token of the next one.
+
     pub fn subgraph(
         &self,
         corpus_name: &str,